% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationalpatterns.R
\name{sig_plot_mbs}
\alias{sig_plot_mbs}
\title{Plot MBS Mutation Characteristics}
\usage{
sig_plot_mbs(mbs_counts, same_y = TRUE)
}
\arguments{
\item{mbs_counts}{MBS context counts (matrix with counts by length).}

\item{same_y}{Logical. If TRUE, all facets have the same y-axis scale. Default is TRUE.}
}
\value{
A ggplot2 object showing MBS contexts.
}
\description{
Plots MBS (Multi-Base Substitution) mutation characteristics.
}
