% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationalpatterns.R
\name{sig_plot_rainfall}
\alias{sig_plot_rainfall}
\title{Plot Rainfall Plot}
\usage{
sig_plot_rainfall(vcf_gr, ref_genome)
}
\arguments{
\item{vcf_gr}{GRanges containing all mutation types from a single sample.}

\item{ref_genome}{The BSGenome reference genome to use.}
}
\value{
A ggplot2 object representing the rainfall plot, or an informative
placeholder plot if the rainfall plot cannot be generated.
}
\description{
Generates a rainfall plot showing intermutational distances.
}
