% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationalpatterns.R
\name{sig_plot_strand_bias}
\alias{sig_plot_strand_bias}
\title{Plot Strand Bias Analysis}
\usage{
sig_plot_strand_bias(vcf_gr, ref_genome)
}
\arguments{
\item{vcf_gr}{GRanges containing all mutation types from a single sample.}

\item{ref_genome}{The BSGenome reference genome to use.}
}
\value{
A list with four ggplot2 objects:
\itemize{
\item p_transcriptional_bias: transcriptional strand bias plot
\item p_transcriptional_effect: transcriptional strand bias effect size
\item p_replicative_bias: replicative strand bias plot
\item p_replicative_effect: replicative strand bias effect size
}
}
\description{
Plots transcriptional and replicative strand bias analysis.
}
