% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationalpatterns.R
\name{sig_workflow_run}
\alias{sig_workflow_run}
\title{Run MutationalPatterns Workflow}
\usage{
sig_workflow_run(
  vcf,
  sample_nm,
  ref_genome = "hg38",
  outdir,
  rainfall = FALSE,
  strand_bias = FALSE
)
}
\arguments{
\item{vcf}{VCF file with SNVs.}

\item{sample_nm}{Sample name.}

\item{ref_genome}{Human genome assembly string: hg38 (default), hg19 or GRCh37.}

\item{outdir}{Directory path to output all the results to.}

\item{rainfall}{Logical. Whether to generate rainfall plot. Default is FALSE.}

\item{strand_bias}{Logical. Whether to generate strand bias analysis. Default is FALSE.}

\item{predefined_dbs_mbs}{Logical. Whether DBS/MBS variants are predefined in the VCF. Default is TRUE.}
}
\description{
Runs a MutationalPatterns Workflow.
}
\details{
Writes plots and signature contribution tables to \code{outdir}.
}
