% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigrap-package.R
\docType{package}
\name{sigrap-package}
\alias{sigrap}
\alias{sigrap-package}
\title{sigrap: Wrappers for somatic mutation signature analysis tools}
\description{
Wraps functionality from somatic signature analysis tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://umccr.github.io/sigrap/}
  \item \url{https://github.com/umccr/sigrap}
  \item Report bugs at \url{https://github.com/umccr/sigrap/issues}
}

}
\author{
\strong{Maintainer}: Quentin Clayssen \email{quentin.clayssen@gmail.com} (\href{https://orcid.org/0000-0002-2574-073X}{ORCID})

Authors:
\itemize{
  \item Peter Diakumis \email{peterdiakumis@gmail.com} (\href{https://orcid.org/0000-0002-7502-7545}{ORCID})
}

}
\keyword{internal}
