% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationalpatterns.R
\name{sig_count_mbs}
\alias{sig_count_mbs}
\title{Count mbs Contexts}
\usage{
sig_count_mbs(vcf_gr)
}
\arguments{
\item{vcf_gr}{GRanges containing all mutation types from a single sample.}

\item{predefined_dbs_mbs}{Logical. Whether MBS variants are predefined in the VCF
(TRUE) or need to be detected from adjacent SNVs (FALSE). Default is TRUE.}
}
\value{
A matrix containing the number of MBSs per COSMIC context per gr.
}
\description{
Counts mbs (Multi-Base Substitution) Contexts.
}
